refactor(jellyfish-api-core): change imports back to using loan
