+++ packages/jellyfish-api-core/__tests__/category/loan/takeLoan.test.ts
@@ -3,7 +3,7 @@ import { GenesisKeys } from '@defichain/testcontainers'
 import BigNumber from 'bignumber.js'
 import { TestingGroup } from '@defichain/jellyfish-testing'
 import { RpcApiError } from '@defichain/jellyfish-api-core'
-import { VaultActive, VaultLiquidation } from '../../../src/category/vault'
+import { VaultActive, VaultLiquidation } from '../../../src/category/loan'
 
 const tGroup = TestingGroup.create(2, i => new LoanMasterNodeRegTestContainer(GenesisKeys[i]))
 const alice = tGroup.get(0)
